+++ src/main/java/com/monits/anviz/model/CheckIn.java
@@ -4,7 +4,7 @@ import org.joda.time.DateTime;
 
 public class CheckIn extends Action {
 
-	public CheckIn(short id, DateTime time) {
+	public CheckIn(long id, DateTime time) {
 		super(id, time);
 	}
